backend/requirements.txt
flask
requests
Pillow
boto3
